Set UID and proper directory permission for AppNet container (#3279)
