+++ agent/engine/service_connect/manager_linux_prvileged_test.go
@@ -1,0 +1,10 @@
+//go:build linux && sudo_unit
+// +build linux,sudo_unit
+
+package serviceconnect
+
+import "testing"
+
+func TestAgentContainerModificationsForServiceConnect_Privileged(t *testing.T) {
+	testAgentContainerModificationsForServiceConnect(t, true)
+}
