+++ agent/engine/service_connect/manager_linux_test.go
@@ -17,51 +17,13 @@
 package serviceconnect
 
 import (
-	"encoding/json"
-	"fmt"
-	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/aws/amazon-ecs-agent/agent/api/serviceconnect"
 
 	apicontainer "github.com/aws/amazon-ecs-agent/agent/api/container"
-	apicontainerstatus "github.com/aws/amazon-ecs-agent/agent/api/container/status"
-	apieni "github.com/aws/amazon-ecs-agent/agent/api/eni"
-	"github.com/aws/amazon-ecs-agent/agent/api/task"
-	apitask "github.com/aws/amazon-ecs-agent/agent/api/task"
-	"github.com/aws/amazon-ecs-agent/agent/config"
-	"github.com/aws/amazon-ecs-agent/agent/engine/testdata"
-	"github.com/aws/aws-sdk-go/aws"
 	dockercontainer "github.com/docker/docker/api/types/container"
 	"github.com/stretchr/testify/assert"
-)
-
-const (
-	ipv4        = "10.0.0.1"
-	gatewayIPv4 = "10.0.0.2/20"
-	mac         = "1.2.3.4"
-	ipv6        = "f0:234:23"
-)
-
-var (
-	cfg     config.Config
-	mockENI = &apieni.ENI{
-		ID: "eni-id",
-		IPV4Addresses: []*apieni.ENIIPV4Address{
-			{
-				Primary: true,
-				Address: ipv4,
-			},
-		},
-		MacAddress: mac,
-		IPV6Addresses: []*apieni.ENIIPV6Address{
-			{
-				Address: ipv6,
-			},
-		},
-		SubnetGatewayIPV4Address: gatewayIPv4,
-	}
 )
 
 func TestDNSConfigToDockerExtraHostsFormat(t *testing.T) {
@@ -95,57 +57,6 @@ func TestDNSConfigToDockerExtraHostsFormat(t *testing.T) {
 		res := DNSConfigToDockerExtraHostsFormat(tc.dnsConfigs)
 		assert.Equal(t, tc.expectedRestult, res, "Wrong docker host config ")
 	}
-}
-
-func getAWSVPCTask(t *testing.T) (*apitask.Task, *apicontainer.Container, *apicontainer.Container) {
-	sleepTask := testdata.LoadTask("sleep5TwoContainers")
-
-	sleepTask.ServiceConnectConfig = &serviceconnect.Config{
-		ContainerName: "service-connect",
-		DNSConfig: []serviceconnect.DNSConfigEntry{
-			{
-				HostName: "host1.my.corp",
-				Address:  "169.254.1.1",
-			},
-			{
-				HostName: "host1.my.corp",
-				Address:  "ff06::c4",
-			},
-		},
-	}
-	dockerConfig := dockercontainer.Config{
-		Healthcheck: &dockercontainer.HealthConfig{
-			Test:     []string{"echo", "ok"},
-			Interval: time.Millisecond,
-			Timeout:  time.Second,
-			Retries:  1,
-		},
-	}
-
-	pauseContainer := apicontainer.NewContainerWithSteadyState(apicontainerstatus.ContainerResourcesProvisioned)
-	pauseContainer.TransitionDependenciesMap = make(map[apicontainerstatus.ContainerStatus]apicontainer.TransitionDependencySet)
-	pauseContainer.Name = task.NetworkPauseContainerName
-	pauseContainer.Image = fmt.Sprintf("%s:%s", cfg.PauseContainerImageName, cfg.PauseContainerTag)
-	pauseContainer.Essential = true
-	pauseContainer.Type = apicontainer.ContainerCNIPause
-
-	rawConfig, err := json.Marshal(&dockerConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	serviceConnectContainer := &apicontainer.Container{
-		Name:            sleepTask.ServiceConnectConfig.ContainerName,
-		HealthCheckType: apicontainer.DockerHealthCheckType,
-		DockerConfig: apicontainer.DockerConfig{
-			Config: aws.String(string(rawConfig)),
-		},
-		TransitionDependenciesMap: make(map[apicontainerstatus.ContainerStatus]apicontainer.TransitionDependencySet),
-	}
-	sleepTask.Containers = append(sleepTask.Containers, serviceConnectContainer)
-
-	// Add eni information to the task so the task can add dependency of pause container
-	sleepTask.AddTaskENI(mockENI)
-	return sleepTask, pauseContainer, serviceConnectContainer
 }
 
 func TestPauseContainerModificationsForServiceConnect(t *testing.T) {
@@ -188,72 +99,6 @@ func TestPauseContainerModificationsForServiceConnect(t *testing.T) {
 	}
 }
 
-func TestAgentContainerModificationsForServiceConnect(t *testing.T) {
-	scTask, _, serviceConnectContainer := getAWSVPCTask(t)
-
-	tempDir := t.TempDir()
-	expectedBinds := []string{
-		fmt.Sprintf("%s/status/%s:%s", tempDir, scTask.GetID(), "/some/other/run"),
-		fmt.Sprintf("%s:%s", tempDir, "/not/var/run"),
-	}
-	expectedENVs := map[string]string{
-		"ReLaYgOeShErE":                 "unix:///not/var/run/relay_file_of_holiness",
-		"StAtUsGoEsHeRe":                "/some/other/run/status_file_of_holiness",
-		"APPNET_AGENT_ADMIN_MODE":       "uds",
-		"ENVOY_ENABLE_IAM_AUTH_FOR_XDS": "0",
-	}
-
-	type testCase struct {
-		name          string
-		container     *apicontainer.Container
-		expectedENV   map[string]string
-		expectedBinds []string
-	}
-	testcases := []testCase{
-		{
-			name:          "Service connect container has extra binds/ENV",
-			container:     serviceConnectContainer,
-			expectedENV:   expectedENVs,
-			expectedBinds: expectedBinds,
-		},
-	}
-	// Add test cases for other containers expecting no modifications
-	for _, container := range scTask.Containers {
-		if container != serviceConnectContainer {
-			testcases = append(testcases, testCase{name: container.Name, container: container, expectedENV: map[string]string{}})
-		}
-	}
-	scManager := &manager{
-		relayPathContainer:  "/not/var/run",
-		relayPathHost:       tempDir,
-		relayFileName:       "relay_file_of_holiness",
-		relayENV:            "ReLaYgOeShErE",
-		statusPathContainer: "/some/other/run",
-		statusPathHostRoot:  filepath.Join(tempDir, "status"),
-		statusFileName:      "status_file_of_holiness",
-		statusENV:           "StAtUsGoEsHeRe",
-		adminStatsRequest:   "/give?stats",
-		adminDrainRequest:   "/do?drain",
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.name, func(t *testing.T) {
-			hostConfig := &dockercontainer.HostConfig{}
-			err := scManager.AugmentTaskContainer(scTask, tc.container, hostConfig)
-			if err != nil {
-				t.Fatal(err)
-			}
-			assert.Equal(t, tc.expectedBinds, hostConfig.Binds)
-			assert.Equal(t, tc.expectedENV, tc.container.Environment)
-
-		})
-	}
-	assert.Equal(t, scTask.ServiceConnectConfig.RuntimeConfig.AdminSocketPath, fmt.Sprintf("%s/status/%s/%s", tempDir, scTask.GetID(), "status_file_of_holiness"))
-	assert.Equal(t, scTask.ServiceConnectConfig.RuntimeConfig.StatsRequest, "/give?stats")
-	assert.Equal(t, scTask.ServiceConnectConfig.RuntimeConfig.DrainRequest, "/do?drain")
-
-	config := scTask.GetServiceConnectRuntimeConfig()
-	assert.Equal(t, config.AdminSocketPath, fmt.Sprintf("%s/status/%s/%s", tempDir, scTask.GetID(), "status_file_of_holiness"))
-	assert.Equal(t, config.StatsRequest, "/give?stats")
-	assert.Equal(t, config.DrainRequest, "/do?drain")
+func TestAgentContainerModificationsForServiceConnect_NonPrivileged(t *testing.T) {
+	testAgentContainerModificationsForServiceConnect(t, false)
 }
